+++ src/__init__.py
@@ -2,5 +2,5 @@
 metadata for the package
 """
 
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 __author__ = "Edmundo Garcia: ingega@gmail.com"